+++ src/middleware/error-handler.js
@@ -1,9 +1,16 @@
 'use strict';
 
+var errors = require('../lib/errors');
+
 /**
  * Express route error handler: send back JSON error responses
  */
 module.exports = function errorHandler(err, req, res, next) {
+    // Only log unexpected errors
+    if (!(err instanceof errors.HttpError)) {
+        req.log.error(err, 'error');
+    }
+
     // Set optional headers
     if (err.headers) {
         res.set(err.headers);
@@ -19,7 +26,7 @@ module.exports = function errorHandler(err, req, res, next) {
     response.code = err.appCode || status;
 
     // Additional field error messages
-    if (err.errors) { response.errors = err.errors;}
+    if (err.errors) { response.errors = err.errors; }
 
     res.status(status).json(response);
 };