+++ src/middleware/error-handler.spec.js
@@ -10,7 +10,7 @@ describe('Error Handler', function() {
 
     before(function() {
         mockery.enable({
-            //warnOnUnregistered: false,
+            warnOnUnregistered: false,
             useCleanCache: true
         });
     });
@@ -33,6 +33,7 @@ describe('Error Handler', function() {
 
     it('should respond with default 500 error', function() {
         var err = {message: 'ERROR_MESSAGE'};
+        var req = {log: {error: sinon.spy()}};
         var res = {
             set: sinon.spy(),
             status: sinon.stub(),
@@ -40,7 +41,7 @@ describe('Error Handler', function() {
         };
         res.status.returns(res);
         var next = sinon.spy();
-        errorHandler(err, null, res, next);
+        errorHandler(err, req, res, next);
         expect(res.set.called).to.be.false;
         expect(res.status.calledWith(500)).to.be.true;
         expect(next.called).to.be.false;
@@ -49,6 +50,7 @@ describe('Error Handler', function() {
 
     it('should set optional headers', function() {
         var err = {message: 'ERROR_MESSAGE', headers: {test: 'TEST_HEADER'}};
+        var req = {log: {error: sinon.spy()}};
         var res = {
             set: sinon.spy(),
             status: sinon.stub(),
@@ -56,12 +58,13 @@ describe('Error Handler', function() {
         };
         res.status.returns(res);
         var next = sinon.spy();
-        errorHandler(err, null, res, next);
+        errorHandler(err, req, res, next);
         expect(res.set.calledWith({test: 'TEST_HEADER'})).to.be.true;
     });
 
     it('should set status and app codes', function() {
         var err = {message: 'ERROR_MESSAGE', statusCode: 404, appCode: 1337};
+        var req = {log: {error: sinon.spy()}};
         var res = {
             set: sinon.spy(),
             status: sinon.stub(),
@@ -69,13 +72,14 @@ describe('Error Handler', function() {
         };
         res.status.returns(res);
         var next = sinon.spy();
-        errorHandler(err, null, res, next);
+        errorHandler(err, req, res, next);
         expect(res.status.calledWith(404)).to.be.true;
         expect(res.json.calledWith({message: 'ERROR_MESSAGE', code: 1337}));
     });
 
     it('should set extra field errors', function() {
         var err = {message: 'ERROR_MESSAGE', errors: []};
+        var req = {log: {error: sinon.spy()}};
         var res = {
             set: sinon.spy(),
             status: sinon.stub(),
@@ -83,7 +87,7 @@ describe('Error Handler', function() {
         };
         res.status.returns(res);
         var next = sinon.spy();
-        errorHandler(err, null, res, next);
+        errorHandler(err, req, res, next);
         expect(res.json.calledWith({message: 'ERROR_MESSAGE', code: 500, errors: []}));
     });
 });