+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.0.2 / 2015-06-16
+==================
+
+- update: Support all HTTP Error types - Dynamically creates all error types from HTTP codes
+- update: Log unhandled errors
+
 0.0.1
 =====